requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
sentence-transformers
faiss-cpu
openai
pytesseract
camelot-py
shap
